Show event information on click on event card
Use fetch to send a GET request to the server to render an ejs template
of the event info. The response contains the html as text. We use
eventContainer.innerHTML = response.text() to replace the content on
the event container.

A better approach would use a bundler and send the event information
as json and use it to render the template directly on the client side.