Enable browser history on events page
After replacing the event info on click on a particular event-card,
We append a query to the url and push it to the history. To avoid the
page to refresh completely on a popstate event, we reuse the same
function that fetches the information of the event on click on a
event-card. In this case we replace the state in the history instead.